backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.7.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.10.4
pydantic-settings==2.7.1
psutil==6.1.1
python-multipart==0.0.20
websockets==14.1
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
